refactor(ui): Move the modal cluster from ui/ to custom/modal/
Changes:
- git mv modal-shell, modal-shell-footer, unsaved-indicator, confirm-dialog to components/custom/modal/
- Update all import paths across the app

components/ui/ is meant for shadcn primitives only; these are app-authored compositions.
The remaining app-authored files in ui/ are a follow-up.

Notes:
- Review finding #54 (modal cluster)
