+++ frontend/src/app/(user)/digital-cards/components/forms/digital-card-modal.tsx
@@ -6,7 +6,7 @@ import { zodResolver } from "@hookform/resolvers/zod";
 
 import { Save } from "lucide-react";
 
-import { ModalShell } from "@/components/ui/modal-shell";
+import { ModalShell } from "@/components/custom/modal/modal-shell";
 import { Form } from "@/components/ui/form";
 import { Skeleton } from "@/components/ui/skeleton";
 import DigitalCardFields from "@/app/(user)/digital-cards/components/forms/digital-card-fields";
