+++ frontend/src/app/dashboard/components/admin-contact-detail.tsx
@@ -2,7 +2,7 @@
 
 import { Mail } from "lucide-react";
 
-import { ModalShell } from "@/components/ui/modal-shell";
+import { ModalShell } from "@/components/custom/modal/modal-shell";
 import { Button } from "@/components/ui/button";
 import { formatDate } from "@/utils/strings";
 import { ContactMessageDto } from "@/lib/api/types";
