+++ frontend/src/app/dashboard/components/admin-users-table.tsx
@@ -23,7 +23,7 @@ import {
 } from "@/components/ui/select";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
-import { ConfirmDialog } from "@/components/ui/confirm-dialog";
+import { ConfirmDialog } from "@/components/custom/modal/confirm-dialog";
 import { adminService } from "@/lib/api";
 import {
   AccountType,
