+++ frontend/src/components/custom/contact/contact-modal.tsx
@@ -7,7 +7,7 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { toast } from "sonner";
 import { Send } from "lucide-react";
 
-import { ModalShell } from "@/components/ui/modal-shell";
+import { ModalShell } from "@/components/custom/modal/modal-shell";
 import { Form } from "@/components/ui/form";
 import ContactChannels from "@/components/custom/contact/contact-channels";
 import ContactFormFields from "@/components/custom/contact/contact-form-fields";
