+++ frontend/src/components/custom/settings/security/components/account-actions.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import { LogOut, MonitorSmartphone, Trash2, TriangleAlert } from "lucide-react";
 
 import { Button } from "@/components/ui/button";
-import { ConfirmDialog } from "@/components/ui/confirm-dialog";
+import { ConfirmDialog } from "@/components/custom/modal/confirm-dialog";
 import SettingRow from "@/components/custom/settings/ui/setting-row";
 import { useSecurityActions } from "@/components/custom/settings/security/security-actions";
